refactoring: Layout was renamed to AppLayout. Now Applayout throw Outlet to Main layout instead using Outlet independently.
