+++ src/components/App.js
@@ -2,7 +2,7 @@ import React, { useEffect } from 'react';
 import Login from '../screens/Login/Login';
 import { Routes, Route, } from 'react-router-dom';
 import RequireAuth from '../hoc/RequireAuth';
-import Layout from '../layouts/Layout/Layout';
+import AppLayout from '../layouts/AppLayout/AppLayout';
 import Home from '../screens/Home/Home';
 import People from '../screens/People/People';
 import Settings from '../screens/Settings/Settings';
@@ -28,7 +28,7 @@ function App() {
       <Routes>
         <Route path={ROUTES.initial} element={
           <RequireAuth>
-            <Layout />
+            <AppLayout />
           </RequireAuth>
         }>
           <Route index element={<Home />} />
