+++ src/layouts/AppLayout/AppLayout.jsx
@@ -1,18 +1,21 @@
-import styles from './Layout.module.scss';
+import { Outlet } from "react-router-dom";
+import styles from './AppLayout.module.scss';
 import Header from "../Header/Header";
 import Main from '../Main/Main';
 import Footer from '../Footer/Footer';
 
-const Layout = () => {
+const AppLayout = () => {
   return (
     <div className={styles['main-wrapper']}>
       <div className={styles['main-wrapper__content']}>
         <Header />
-        <Main />
+        <Main>
+          <Outlet />
+        </Main>
         <Footer />
       </div>
     </div>
   )
 }
 
-export default Layout;
+export default AppLayout;
